Cleaned up requirements
The requirements files were starting to require dual entry, and
for sure, the requirements.txt file was not being used. This project
currently relies on the dist-requirements.txt for installation. This
pull makes it so that top-level dependencies are added in
requirements.txt and testing tools are added to
test-requirements.txt.