Use nsenter to process events in the agent-instance
Previously we would make a HTTP call to enter the container to run scripts.
Now that the python-agent can run in the PID namespace of the host we are
now able to just directly enter the namespace using nsenter.  "docker exec"
is not used because there is issues with properly handling input/output, for
example flushing and closing the input stream.