add support for volume plugins
This change adds the capability to do the equivalent of `docker volume
create` and `docker volume delete`.

volume create will happen int two scenarios:

1. If we receive an explicit storage.volume.activate event
2. If we receive a compute.instance.create event that contains volume
data in the `volumesFromDataVolumeMounts` field. That field is an array
of volumes that should be on the host for the container. We iterate
through it and create a volume appropriately.