Added new options for configuring cadvisor
This PR sets a new housekeeping interval to 5s to reduce
the load cAdvisor puts on a host. It also allows the user to
override that setting by adding `-e CATTLE_CADVISOR_INTERVAL=n` for
users that would like to really reduce its impact can set it to
very high values.

Also added, is a setting for `CATTLE_CADVISOR_OPTS` where users can
add options to the cAdvisor command line. This enables more control
for the user around how cAdvisor behaves. It also opens the door for
adding remote storage and metrics.