Name container as r-{name} and not uuid
We would ideally like to name containers what the user has specified but
there are concerns right now about /etc/hosts management that Docker
does overriding our DNS based stuff.  We have compromised right now in
naming containers r-{name}.  If we find the name is not unique or the
container name has invalid characters, we fall back to UUID for the
name.